# Implement engine="thread" in ChunkRecordingExecutor