… doc/get_started/quickstart.rst
… src/spikeinterface/core/tests/test_globals.py
… src/spikeinterface/core/tests/test_job_tools.py
… src/spikeinterface/core/tests/test_waveform_tools.py
… src/spikeinterface/core/waveform_tools.py
… src/spikeinterface/postprocessing/principal_component.py
… src/spikeinterface/postprocessing/tests/test_principal_component.py
… src/spikeinterface/qualitymetrics/pca_metrics.py
… src/spikeinterface/qualitymetrics/tests/conftest.py
… src/spikeinterface/qualitymetrics/tests/test_pca_metrics.py
… src/spikeinterface/sortingcomponents/clustering/merge.py
… src/spikeinterface/sortingcomponents/clustering/split.py