@h-mayorquin @zm711 @alejoe91 
I wanted to do this since a long time : implement **thread for ChunkRecordingExecutor**.

Now we can do :
```python 
# this use threads
job_kwargs = dict(chunk_duration="1s", pool_engine="thread", n_jobs=8, progress_bar=True)
# this use processes
job_kwargs = dict(chunk_duration="1s", pool_engine="process", n_jobs=8, progress_bar=True)
```

Maybe this will help a lot windows user and will avoid use "spawn" for multiprocessing (which is too slow at startup).
My bet is that thread will have IO (read data from disk) and computation withou the GIL without computing lib, so it should good enought for parralell computing. Lets see.

Here a very first as a proof of concept. (Without any real test).

TODO:
* [X] add engine thread
* [X] add worker_index mechanism in job_tools.py (this will be usefull for multi GPU)
* [x] rename `max_threads_per_process` to  `max_threads_per_worker`. See note.
* [ ] discuss renaming n_jobs to `n_worker` or `num_worker` or `max_worker` (Alessio will be unhappy)
* [x] make a `get_best_job_kwargs()` that will be platform dependant.

When using engine thread we will need to disambigu the concept of worker that can be thread and `max_threads_per_worker` which related to the hidden thread use by computing lib like numpy/scipy/blas/lapcak/sklearn.


EDIT:
  * on linux thread are a bit slower than process with fork but not so much.


Important notes:
many `ProcessPoolExecutor` are hard coded in several places (principal_components, split, merge...)
we should also do this but this will be a separated PR. because theses parallelisation are not on the recording axis.
Note that tricky stuff between loops, thread and process is initialzing variable to a private dict per worker. This is done in `ChunkRecordingExecutor` but this should be propagated for other use cases that do not need recording slices.