@zm711 2024-11-20
> @zm711 : if you have time could you make some test using windows ?

Yep. Will do once tests pass :)