@alejoe91 2024-11-20
failing tests are really weird. The arrays that are not equal look exactly the same from the values