@zm711 2024-11-20
Okay with default settings on this branch using the `generate_ground_truth_recording`
```
n_jobs=1
67 ms ± 4.35 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
n_jobs=2
47.9 s ± 971 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# explicitly setting pool_engine=='thread'
same as n_jobs=2
```