@zm711 2024-11-22
I forgot.... easy testing of `save` is not really possible on Windows due to the inability to overwrite files still open in the process so I can't use timeit. I will need to try to write a script. I can try detect_peaks though.