@samuelgarcia 2024-11-22
Salut Zach,
thanks a gain for your time.

theses results are a bit disapointing... I would except better speedup.
Do you have many core  on this windows ?
I will also do some tests.