@yger 2024-11-27
@samuelgarcia . i'm testing it and i have a bug

    SorterClass._run_from_folder(sorter_output_folder, sorter_params, verbose)
  File "/media/cure/Secondary/pierre/softwares/spikeinterface/src/spikeinterface/sorters/internal/spyking_circus2.py", line 190, in _run_from_folder
    prototype = get_prototype(
                ^^^^^^^^^^^^^^
  File "/media/cure/Secondary/pierre/softwares/spikeinterface/src/spikeinterface/sorters/internal/spyking_circus2.py", line 428, in get_prototype
    res = detect_peaks(
          ^^^^^^^^^^^^^
  File "/media/cure/Secondary/pierre/softwares/spikeinterface/src/spikeinterface/sortingcomponents/peak_detection.py", line 133, in detect_peaks
    outs = run_node_pipeline(
           ^^^^^^^^^^^^^^^^^^
  File "/media/cure/Secondary/pierre/softwares/spikeinterface/src/spikeinterface/core/node_pipeline.py", line 586, in run_node_pipeline
    processor.run(recording_slices=recording_slices)
  File "/media/cure/Secondary/pierre/softwares/spikeinterface/src/spikeinterface/core/job_tools.py", line 538, in run
    recording_slices2 = [(thread_local_data, ) + args for args in recording_slices]
                         ~~~~~~~~~~~~~~~~~~~~~~^~~~~~
TypeError: unsupported operand type(s) for +: '_thread._local' and 'int'