@chrishalcrow 2024-11-27
Hello, just some quick benchmarking on a M4 Mac, doing `detect_peaks` on a 96 channel, 20 minute recording (just once per set-up)

THIS PR:
n_jobs = 1 : 22s
n_jobs = 4, "thread" : 26s
n_jobs = 4, "spawn" : 12s

CURRENT MAIN:
n_jobs = 1 : 26s
n_jobs = 4 : 11s

So the "thread" option doesn't seem to be working as expected for this system.